+++ tests/functional/L0_decoupled_mode/test.py
@@ -16,6 +16,7 @@
 
 METADATA = {
     "image_name": "nvcr.io/nvidia/pytorch:{TEST_CONTAINER_VERSION}-py3",
+    "platforms": ["amd64", "arm64"],
 }