+++ tests/functional/L0_example_use_parameters_and_headers/test.py
@@ -31,6 +31,7 @@ from tests.utils import (
 LOGGER = logging.getLogger((__package__ or "main").split(".")[-1])
 METADATA = {
     "image_name": "nvcr.io/nvidia/pytorch:{TEST_CONTAINER_VERSION}-py3",
+    "platforms": ["amd64", "arm64"],
 }