+++ tests/functional/L0_resources_released_after_signals/server.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,9 +22,6 @@ from tests.functional.common.models import ADD_SUB_PYTHON_MODEL
 from tests.utils import find_free_port
 
 LOGGER = logging.getLogger((__package__ or "main").split(".")[-1])
-METADATA = {
-    "image_name": "nvcr.io/nvidia/pytorch:{TEST_CONTAINER_VERSION}-py3",
-}
 
 
 def main():