+++ tests/functional/L0_resources_released_after_signals/test.py
@@ -35,6 +35,7 @@ LOGGER = logging.getLogger((__package__ or "main").split(".")[-1])
 
 METADATA = {
     "image_name": "nvcr.io/nvidia/pytorch:{TEST_CONTAINER_VERSION}-py3",
+    "platforms": ["amd64", "arm64"],
 }